+++ App/src/App.scss
@@ -51,9 +51,4 @@
 
 	overflow-y: scroll;
 
-	& > div:not(#navbar) {
-
-
-	}
-
 }
