+++ App/src/components/mealplan/Mealplan.scss
@@ -57,7 +57,7 @@
 				justify-self: center;
 				align-self: center;
 
-				background-image: url(../assets/boubbles.jpg);
+				background-image: url(../../assets/boubbles.jpg);
 				background-position: center;
 				background-size: 500%;
