+++ App/src/components/navbar/Navbar.scss
