+++ App/src/components/recipes/Recipes.scss
@@ -54,7 +54,7 @@
 			justify-self: center;
 			align-self: center;
 
-			background-image: url(../assets/boubbles.jpg);
+			background-image: url(../../assets/boubbles.jpg);
 			background-position: center;
 			background-size: 500%;
