+++ App/src/components/selectionArea/SelectionArea.scss
