+++ App/src/components/selectionArea/SelectionArea.tsx
